+++ Dependencies/napi/packages.config
