+++ Dependencies/xr/Source/XrPlatformCommon.h
