+++ Plugins/NativeEngine/Source/BgfxCallback.cpp
