+++ Plugins/NativeEngine/Source/ShaderCompiler.cpp
