+++ Plugins/XMLHttpRequest/Source/XMLHttpRequest.cpp
