+++ Plugins/XMLHttpRequest/Source/XMLHttpRequestAndroid.cpp
