+++ Plugins/XMLHttpRequest/Source/XMLHttpRequestApple.cpp
