+++ Plugins/XMLHttpRequest/Source/XMLHttpRequestUWP.cpp
