+++ Plugins/XMLHttpRequest/Source/XMLHttpRequestWin32.cpp
